libraries/history-tree: Allow jumping back in history.
This is quite a debatable one.

One the one hand, it makes sense to jump back in
history when one reaches some hub-like URL. They want to get back to the
crossroads they've  already seen, and they expect that the paths they followed
already are still there.

On the other hand, one may see things as a linear narrative, in which there's no
going back. They might still benefit from the forward suggestions, but their
mental map could be broken because of this back-stitching.